[CHORE] : activity안에 recyclerview를 먹/갈/할 fragment로 나누었음, 서버나오면 다시 작업 예정
