+++ app/src/main/res/layout/activity_thunder_list.xml
@@ -123,6 +123,15 @@
 
         </LinearLayout>
 
+        <androidx.viewpager2.widget.ViewPager2
+            android:id="@+id/vp_thunderlist_container"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_marginHorizontal="20dp"
+            android:layout_marginTop="24dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/ll_thunder_align_container" />
+
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_thunderlist_container"
             android:layout_width="match_parent"
